feat: improve wrappers
- added wrapers for parented transform operations in 2D & 3D
- converted components static methods into regular instance methods
- created a static "ctor" for cameras

+++ src/cat/gfx/sprite_atlas.cpp
@@ -22,7 +22,7 @@ TextureUV SpriteAtlas::get_uv(const glm::ivec2& sprite_index) const
     return uv;
 }
 
-cSprite SpriteAtlas::get_sprite(const glm::ivec2& index, u32 z_index) const
+cSprite SpriteAtlas::get_sprite(const glm::ivec2& index, u8 z_index) const
 {
     return cSprite
     {
